% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMutils.R
\name{boot.se}
\alias{boot.se}
\title{Compute bootstrap standard errors}
\usage{
boot.se(fit, boot, probs = c(0.025, 0.975), z0 = 1.96, b0 = 0)
}
\arguments{
\item{fit}{An object of class \code{glm} or \code{lm}.}

\item{boot}{A bootstrapped model object (see the examples section).}

\item{probs}{Bootstrap intervals (default = c(0.025, 0.975)).}

\item{z0}{z-score under null hypothesis (default = 1.96).}

\item{b0}{Effect size under null hypothesis (default = 0).}

\item{...}{Currently ignored.}
}
\value{
A data.frame of 9 columns:
\enumerate{
\item "Variable", variable name;
\item "Estimate", parameter (effect size) estimate;
\item "se.boot", bootstrap standard error;
\item "lower", confidence interval lower bound;
\item "upper", confidence interval upper bound;
\item "conf.level", confidence level;
\item "method", estimation method;
\item "z", z-score = (estimate - b0)/SE;
\item "P", 2-sided p-value; i.e., 2*pnorm(-abs(z)).
}
}
\description{
Compute bootstrap standard errors (SE) for a given 
   (generalized) linear model.
}
\examples{

# Dichotomized dataset creation
x <- dichotomize(mpm.us, asFactor = TRUE)

# Model specification
model <- formula(paste0(c("y ~ shortAxis + cortical + hilum + ",
                          "inflammatoryStroma + extracapsularSpread + ",
                          "ecostructure + FID + VFL + corticalThickening + ",
                          "vascularPattern + CMID + shape + grouping + ",
                          "colorScore"), collapse = ""))

# Binomial model fitting (MLE)
fit <- glm(model, data = x, family = "binomial")

# Binomial model fitting
n.reps <- 100
boot <- do(n.reps) * coef(glm(model, data = resample(x), family = "binomial"))

# Bootstrap SE calculation
SE <- boot.se(fit, boot)
print(SE)

}
\references{
Pruim R, Kaplan DT, Horton NJ (2017). The mosaic Package: Helping 
Students to 'Think with Data' Using R. The R Journal, 9(1), 77–102. 
<https://journal.r-project.org/archive/2017/RJ-2017-024/index.html>
}
\seealso{
See \code{\link[morphonode]{p.boot}} for performance indices 
   bootstrap confidence intervals.
   See also \code{\link[mosaic]{do}} for resampling.
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
