% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMutils.R
\name{brier}
\alias{brier}
\title{Compute Brier scores}
\usage{
brier(data, status, k = 5, method = "CV", ...)
}
\arguments{
\item{data}{An (n, m) matrix or data.frame with no outcome attribute.}

\item{status}{A vector of length n, containing the outcome.}

\item{k}{number of cross-validation iterations (default = 5).}

\item{method}{One of the \code{\link[CMA]{GenerateLearningsets}} 
methods, including: "LOOCV", "CV", "MCCV", and "bootstrap" 
(default = "CV").}

\item{...}{Currently ignored.}
}
\value{
A vector of Brier scores of length n (one value per subject).
}
\description{
Compute Brier scores for a given dataset.
}
\examples{

# Extract a subset of 300 subjects and an outcome vector of length 30 
# from the default simulated dataset

x <- mosaic::sample(mpm.us, 300, replace = FALSE, prob = NULL)
y <- x$y
x <- x[, 2:15]
print(dim(x))
print(length(y))

# Compute brier scores
E <- brier(x, y)
print(quantile(E))

}
\references{
Slawski M, Daumer M, Boulesteix AL. CMA - a comprehensive Bioconductor 
package for supervised classification with high dimensional data. 
BMC Bioinformatics 9, 439 (2008).
<https://doi.org/10.1186/1471-2105-9-439>

Brier GW. Verification of forecasts expressed in terms of probability. 
Monthly Weather Review. 1950;78(1):1-3. 
<https://doi.org/10.1175/1520-0493(1950)078<0001:VOFEIT>2.0.CO;2>
}
\seealso{
\code{\link[morphonode]{loss}}
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
