% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMcore.R
\name{check.rfcdata}
\alias{check.rfcdata}
\title{Check RFC input data}
\usage{
check.rfcdata(x, levels = NULL, ...)
}
\arguments{
\item{x}{An (n, 14) ultrasound features data.frame, where n is the 
number of subjects.}

\item{levels}{A list of length 15, corresponding to the levels of 
each ultrasound variable plus the phenotype (y = {0, 1}). 
Needed for categorical variables (factors) checking; 
for continuous variables, it should assume the nominal value of 0. 
If NULL (default), the \code{\link[morphonode]{mpm.levels}} object 
will be used.}

\item{...}{Currently ignored.}
}
\value{
A data.frame of ultrasound features with checked data type.
}
\description{
Assign ultasound data types for RFC building and validation.
}
\examples{

# generate a dataset of 500 subjects
x <- mosaic::sample(mpm.us, 500, replace = FALSE, prob = NULL)
x <- x[, 2:16]

# Assign the correct data types for RFC building and validation
x <- check.rfcdata(x)
summary(x)

}
\seealso{
\code{\link[morphonode]{buildPredictor}}.
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
