% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMcore.R
\name{dichotomize}
\alias{dichotomize}
\title{Generates a dichotomous ultrasound feature data.frame}
\usage{
dichotomize(
  x,
  short = 8,
  cortical = 2,
  ist = 1,
  ecs = 1,
  hab = 0,
  eco = 1,
  vp = c(1, 2, 3),
  vfl = c(2, 3, 4),
  ct = c(2),
  fid = c(1, 2, 3),
  cmid = c(2, 3, 4),
  shape = c(3),
  cs = 3,
  grouping = c(2, 3),
  asFactor = FALSE,
  ...
)
}
\arguments{
\item{x}{An (n, 14) ultrasound features data.frame, where n is the 
number of subjects.}

\item{short}{Numeric value corresponding to the short axis cutoff 
in millimeters (default = 8).}

\item{cortical}{Numeric value corresponding to the cortical thickness 
cutoff in millimeters (default = 2).}

\item{ist}{Dichotomous value {0, 1} for the presence of inflammatory 
stroma (perinodal hyperechogenic ring; default = 1).}

\item{ecs}{Dichotomous value {0, 1} for the presence of extracapsular 
spread (cortical interruption; default = 1).}

\item{hab}{Dichotomous value {0, 1} for the absence of the hilum 
(nodal core sign; default = 0).}

\item{eco}{Dichotomous value {0, 1} for heterogeneous echogenicity 
(echostructure; default = 1).}

\item{vp}{Categorical value (integers from 0 to 4) associated to a 
high-risk vascular flow architecture pattern (default = c(1, 2, 3)).}

\item{vfl}{Categorical value (integers from 0 to 4) associated to a 
high-risk vascular flow localization (default = c(2, 3, 4)).}

\item{ct}{Categorical value (integers from 0 to 4) associated to a 
high-risk cortical thickening (default = 2).}

\item{fid}{Categorical value (integers from 0 to 3) associated to a 
high-risk focal intranodal deposit (default = c(1, 2, 3)).}

\item{cmid}{Categorical value (integers from 0 to 4) associated to a 
high-risk cortical-medullar interface distortion (default = c(2, 3, 4)).}

\item{shape}{Categorical value (integers from 1 to 3) associated to a 
high-risk shape (default = 3).}

\item{cs}{Ordinal value (integers from 1 to 5) associated to a 
high-risk color score (default = 3).}

\item{grouping}{Categorical value (integers from 1 to 3) associated 
to a high-risk grouping (default = c(2, 3)).}

\item{asFactor}{Logical value. If TRUE, data.frame columns are converted 
to factors (default = FALSE).}

\item{...}{Currently ignored.}
}
\value{
An ultrasound profile with imputed missing values.
}
\description{
Convert an ultrasound feature data.frame (each row is 
   an ultrasound vector) into a dichotomous data.frame. 
   Defalut dichotomization cutoffs are computed as described in 
   Fragomeni et al. (2022).
}
\details{
This function is internally used to estimate the malignancy 
   risk through the robust binomial model (Morphonode-RBM).
   Dichotomization is performed to avoid the extremely low frequancy 
   of some levels in categorical variables.
}
\examples{

# Create a dichotomous version of subjects with metastatic signature, 
# from the default simulated dataset.

M <- dichotomize(mpm.us[mpm.us$signature == "MET", 2:15])
print(head(M))

}
\references{
Fragomeni SM, Moro F, Palluzzi F, Mascilini F, Rufini V, Collarino A, 
Inzani F, Giacò L, Scambia G, Testa AC, Garganese G (2022). 
Evaluating the risk of inguinal lymph node metastases before surgery 
using the Morphonode Predictive Model: a prospective diagnostic study. 
Ultrasound xx Xxxxxxxxxx xxx Xxxxxxxxxx. 00(0):000-000.
<https://doi.org/00.0000/00000000000000000000>
}
\seealso{
See \code{\link[morphonode]{uss}} for metastatic risk 
   signature detection (Morphonode-DT module).
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
