% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMutils.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean distance}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{...}{Currently ignored.}
}
\value{
A numeric value corresponding to the euclidean distrance.
}
\description{
Compute the euclidean distance between two vectors.
}
\examples{

# Sample two random ultrasound profiles from the default dataset
x <- sample(mpm.us, 1, replace = FALSE, prob = NULL)
x <- as.matrix(x[, 2:15])
y <- sample(mpm.us, 1, replace = FALSE, prob = NULL)
y <- as.matrix(y[, 2:15])

# Compute the euclidean distance
d <- euclidean(x, y)
print(d)

}
\seealso{
\code{\link[morphonode]{similarity}}, 
   \code{\link[morphonode]{jaccard}}
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
