% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPM_env.R
\docType{data}
\name{mpm.env}
\alias{mpm.env}
\title{Morphonode Predictive Model environment}
\format{
"mpm.env" is a data.frame of 14 rows (one for each ultrasound feature) 
   and 5 columns, including: descriptions (variable name), units (unit 
   of measure for continuous variables, and levels description for 
   categorical variables), variables (variable name within the morphonode 
   package), values (possible values for each variable), basal (basal 
   value for each variable).
}
\usage{
mpm.env
}
\description{
Ultrasound features description.
}
\examples{

# Show morphonode environment
mpm.env()

}
