% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPM_levels.R
\docType{data}
\name{mpm.levels}
\alias{mpm.levels}
\title{Ultrasound features levels}
\format{
"mpm.levels" is a list of 14 vectors, each containing the levels of 
an ultrasound feature. Continuous features (i.e., short axis and cortical 
thickness) have the nominal value of 0. Levels are used by predictive 
functions to switch from numeric to factor classes.
}
\usage{
mpm.levels
}
\description{
A list of 14 elements containing the levels of each 
   ultrasound feature.
}
\examples{

# Generate a simulated malignant ultrasound profile
x <- new.profile(us.simulate(y = 1))

# Prepare the profile for RFC malignancy prediction
u <- set.rfcdata(x, ref = mpm.us[, 2:15], levels = mpm.levels)
print(u)

# Predict subject's phenotype
P <- rfc.predict(u$ultrasound, rfc = mpm.rfc$rfc)
print(P)

}
