% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPM_RBM_sim.R
\docType{data}
\name{mpm.rbm}
\alias{mpm.rbm}
\title{Morphonode default Robust Binomial Model (RBM) endemble}
\format{
"mpm.rbm" is a list of 3 objects:
\enumerate{
\item "coef", a data.frame reporting bootstrap-based estimations:
   ultrasound feature (Variable), log(odds ratio) (Estimate), 
   bootstrap standare errors (se.boot), confidence interval lower 
   bound (lower), confidence interval upper bound (upper), 
   confidence level (conf.level), bootstrap estimation method (method), 
   z-score (z), 2-sided p-value (P);
\item "model", \code{R} formula representing the fitted model;
\item "fit", MLE-based fitted model object of class \code{glm}.
}
}
\usage{
mpm.rbm
}
\description{
Logistic model generated by fitting a simulated dataset 
   of 948 ultrasound profiles (440 malignant and 508 non-malignant), 
   with robust bootstrap standard errors estimation (5000 bootstrap 
   iterations), through the internal morphonode function 
   \code{\link[morphonode]{boot.se}}. The input ultrasound feature 
   dataset was dichotomized before model fitting (see 
   \code{\link[morphonode]{dichotomize}}) to avoid parameter estimation 
   biases due to very low frequency of the levels of some categorical 
   ultrasound features. The RBM is used to estimate the malignancy risk, 
   (R) providing a continuous measure of malignancy (in contrast to the 
   dichotomous prediction of the RFC ensemble). Considering the 
   expected simulated phenotype (y) as the ground truth, two optimal 
   malignancy risk cutoffs were estimated, defining three risk levels: 
   low (R < 0.23), moderate (0.23 <= R <= 0.29), and high (R > 0.29).
}
\examples{

# Create a simulated malignant ultrasound profile
x <- new.profile(simulate.us(y = 1))

# Lauch the Morhonode Predictive Model
u <- us.predict(x)

}
\references{
Fragomeni SM, Moro F, Palluzzi F, Mascilini F, Rufini V, Collarino A, 
Inzani F, Giacò L, Scambia G, Testa AC, Garganese G (2022). 
Evaluating the risk of inguinal lymph node metastases before surgery 
using the Morphonode Predictive Model: a prospective diagnostic study. 
Ultrasound xx Xxxxxxxxxx xxx Xxxxxxxxxx. 00(0):000-000.
<https://doi.org/00.0000/00000000000000000000>
}
