% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPM_ultrasound_sim.R
\docType{data}
\name{mpm.us}
\alias{mpm.us}
\title{Morphonode simulated ultrasound feature dataset}
\format{
"mpm.us" is a data.frame of 948 rows (simulated ultrasound profiles) 
and 18 columns, including: a progressive number used as unique profile 
identifier (ID), 14 ultrasound features used for RFC and RBM building, 
expected simulation phenotype used as ground truth (y = {0: non-malignant, 
1: malignant}), metastatic risk signature associated to each simulated 
ultrasound profile (signature), subject-level prediction error calculated 
as Brier score (E).
}
\usage{
mpm.us
}
\description{
A simulated dataset of 948 ultrasound profiles 
   (440 malignant and 508 non-malignant) used to build the default 
   ensemble random forest classifier (RFC) and the robust binomial 
   model (RBM). This dataset was generated as a 4-fold expansion of 
   the original ultrasound feature dataset of 237 groin samples (75 
   malignant and 162 non-malignant) from Fragomeni et al. (2022), 
   using the morphonode function \code{\link[morphonode]{us.simulate}}.
}
\examples{

# Default simulated dataset phenotype frequencies
print(table(mpm.us$y))

# Default simulated dataset signatures frequencies
print(table(mpm.us$signature))

# Default simulated dataset prediction error quartiles
print(quantile(mpm.us$E))

}
\references{
Fragomeni SM, Moro F, Palluzzi F, Mascilini F, Rufini V, Collarino A, 
Inzani F, Giacò L, Scambia G, Testa AC, Garganese G (2022). 
Evaluating the risk of inguinal lymph node metastases before surgery 
using the Morphonode Predictive Model: a prospective diagnostic study. 
Ultrasound xx Xxxxxxxxxx xxx Xxxxxxxxxx. 00(0):000-000.
<https://doi.org/00.0000/00000000000000000000>
}
