% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMcore.R
\name{new.profile}
\alias{new.profile}
\title{Create a new ultrasound profile}
\usage{
new.profile(u = NULL, ...)
}
\arguments{
\item{u}{A vector of length 14, corresponding to the input ultrasound 
profile. The order of the elements in u must be: 
short axis [numeric], cortical thickness [numeric], 
nodal core sign (hilum) [dichotomous], 
perinodal hyperechogenic ring [dichotomous], 
cortical interruption [dichotomous], echogenicity [dichotomous], 
focal intranodal deposit [categorical], 
vascular flow localization [categorical], 
cortical thickening [categorical], 
vascular flow architecture pattern [categorical], 
cortical-medullar interface distortion [categorical], 
shape [categorical], grouping [categorical], color score [ordinal].
Value -1 can be entered for missing values.
If \code{u = NULL}, a guided interactive prompt will be launched.
In the interactive mode, typing return without entering any value 
will introduce a missing value (this is not possible for short axis 
and cortical thickness, since they are necessary for a reliable 
prediction).}

\item{...}{Currently ignored.}
}
\value{
A list of 2 objects:
\enumerate{
\item "ultrasound", ultrasound features vector;
\item "missing", missing values vector (empty, if no missing values 
                 are found).
}
}
\description{
Creates a new ultrasound profile, used as input for the 
\code{\link[morphonode]{us.predict}} function.
}
\examples{

### Profile with missing data

u <- new.profile(c(10.0, 6.3, 1, 0, 0, 0, -1, 1, 2, 2, 3, -1, -1, -1))
print(u)

### New profile from simulated data

# High metastatic risk profile
u.hmr <- new.profile(simulate.us(signature = "HMR"))
print(u.hmr)

# Low metastatic risk profile
u.lmr <- new.profile(simulate.us(signature = "LMR"))
print(u.lmr)

# Malignant profile
u1 <- new.profile(simulate.us(y = 1))
print(u1)

# Non-malignant profile
u0 <- new.profile(simulate.us(y = 0))
print(u0)

}
\seealso{
See \code{\link[morphonode]{show.mpmenv}} for ultrasound 
variable description.
See also \code{\link[morphonode]{us.predict}} to launch the full 
   Morphonode Predictive Model suite.
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
