% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMutils.R
\name{performance}
\alias{performance}
\title{Predictor performance calculation}
\usage{
performance(obs = NULL, pred = NULL, C = NULL, y = "0,1")
}
\arguments{
\item{obs}{A vector of observed values.}

\item{pred}{A vector of predicted values.}

\item{C}{2x2 contingency table (alternative to \code{obs} and 
\code{pred}). If \code{obs} and \code{pred} are given, the 
contingency table is automatically computed. The contingency 
table has the observed values as rows and the predicted ones as 
columns. By default, true negatives are located at position [1, 1], 
while true positives at [2, 2] (see below).}

\item{y}{Contingenty table orientation. If \code{y = "0,1"} (default), 
true negatives are located at position [1, 1], while true positives 
at position [2, 2]. If \code{y = "1,0"}, these positions are inverted.}

\item{...}{Currently ignored.}
}
\value{
A list of 6 objects:
\enumerate{
\item "ctable", 2x2 contingency table of predicted vs. observed values;
\item "sensitivity", Se = TP/(TP + FN);
\item "specificity", Sp = TN/(TN + FP);
\item "precision", PPV = TP/(TP + FP), 
   also called "positive predictive value";
\item "NPV", NPV = TN/(TN + FN), 
   "negative predictive value";
\item "F1", F1 = 2*Se*PPV/(Se + PPV);
\item "accuracy", (TP + TN)/(TP + TN + FP + FN).
}
}
\description{
Compute sensitivity, speciticity, precision, F1 score, 
   and accuracy for a set of predictions.
}
\examples{

# RBM performance
p <- predict(mpm.rbm$fit, dichotomize(mpm.us[2:15], asFactor = TRUE),
             type = "response")
y.hat <- ifelse(p > 0.5, 1, 0)
rbm <- performance(mpm.us$y, y.hat)

# Compare RBM and RFC F1 scores
print(rbm$F1)
print(mpm.rfc$performance$F1)

}
\seealso{
\code{\link[morphonode]{us.predict}}
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
