% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMutils.R
\name{rfc}
\alias{rfc}
\title{Build a random forest classifier}
\usage{
rfc(data, status, k = 5, method = "CV", ...)
}
\arguments{
\item{data}{An (n, m) matrix or data.frame with no outcome attribute.}

\item{status}{A vector of length n, containing the outcome.}

\item{k}{number of cross-validation iterations (default = 5).}

\item{method}{One of the \code{\link[CMA]{GenerateLearningsets}} 
methods, including: "LOOCV", "CV", "MCCV", and "bootstrap" 
(default = "CV").}

\item{...}{Currently ignored.}
}
\value{
A list of objects of class "cloutput" and 
   "clvarseloutput", respectively.
}
\description{
Wrapper for the \code{\link[CMA]{GenerateLearningsets}} 
   \code{\link[CMA]{classification}} functions.
}
\examples{

# Extract a subset of 300 subjects and an outcome vector of length 30 
# from the default simulated dataset

x <- mosaic::sample(mpm.us, 300, replace = FALSE, prob = NULL)
y <- x$y
x <- x[, 2:15]
dim(x)
length(y)

# Build a 5-fold cross-validation object

CV <- rfc(x, status = y)

# Performances of the first of five predictors

CV1 <- CV[[1]]
P <- performance(obs = CV1@y, pred = CV1@yhat)
print(P)

}
\references{
Slawski M, Daumer M, Boulesteix AL. CMA - a comprehensive Bioconductor 
package for supervised classification with high dimensional data. 
BMC Bioinformatics 9, 439 (2008).
<https://doi.org/10.1186/1471-2105-9-439>
}
\seealso{
See \code{\link[morphonode]{us.predict}} to launch all 
   morphonode modules at once.
   See also \code{\link[morphonode]{topsim}} for a simple similarity 
   search.
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
