% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMcore.R
\name{set.missing}
\alias{set.missing}
\title{Cope with possible missing data in the ultrasound profile}
\usage{
set.missing(
  v,
  ref = NULL,
  levels = NULL,
  con = 1:2,
  cat = 3:14,
  missing = -1,
  na = NA,
  asNumeric = TRUE,
  ...
)
}
\arguments{
\item{v}{An ultrasound profile generated by 
\code{\link[morphonode]{new.profile}}, i.e., a list of two objects, 
containing a vector of length 14 (\code{list$ultasound}), 
corresponding to the input ultrasound profile (14 ultrasound 
variables or "features"), and a vector of missing value indices 
(\code{list$missing}).}

\item{ref}{A data.frame representing the reference dataset. The 
ultrasound profile will be attached to the reference dataset 
before the imputation. This argument is required to impute missing 
features.}

\item{levels}{A list of length 14, corresponding to the levels of 
each ultrasound variable. Needed for categorical variables (factors); 
for continuous variables, it should assume the nominal value of 0. 
The default levels variable \code{mpm.levels} can be used.}

\item{con}{Vector of the indices corresponding to continuous variables 
in the \code{list$ultasound} vector (default = 1:2).}

\item{cat}{Vector of the indices corresponding to categorical variables 
in the \code{list$ultasound} vector (default = 3:14).}

\item{missing}{Value used to mark missing data (default = -1).}

\item{na}{Value used for "not available" data (default = NA). This 
will be used to dubstitute \code{missing} within the ultrasound 
vector before the imputation.}

\item{asNumeric}{Logical value used to convert every value in the 
ultrasound vector to class "numeric". This argument is used only 
if \code{ref = NULL} and \code{levels = NULL}.}

\item{...}{Currently ignored.}
}
\value{
An ultrasound profile with imputed missing values.
}
\description{
This function imputes missing data in the ultrasound 
profile, creating a new profile with imputed missing values.
If no missing values are found, it will simply send a message and 
return the input profile.
}
\details{
Automatic imputation is necessary to improve RFC-based 
   (malignancy prediction) and RBM-based (metastatic risk evaluation) 
   estimations. Imputation is currently forbidden for short axis and 
   cortical thickness (i.e., the first two ultrasound features), since 
   they have a critical role in the prediction, estimation and 
   signature detection processes. Hence, their actual value must be 
   entered for a reliable prediction. Although permitted, the imputation 
   is discouraged for the following three features: 
   nodal core sign (i.e., hilum presence), perinodal hyperechogenic 
   ring (i.e., the presence of inflammatory stroma), and 
   cortical interruption (i.e., extracapsular spread).
   These features define a strongly metastatic profile with possible 
   multiple metastases (i.e., the "MET" signature) that are hardly 
   imputable from the other ultrasound variables.
}
\examples{

# Create an ultrasound profile with missing values

u <- new.profile(c(10.0, 6.3, 1, 0, 0, 0, -1, 1, 2, 2, 3, 1, -1, -1))
print(u)

# Fix missing values with the default simulated dataset as reference
# (ultrasound features only: \code{mpm.us} attributes 2 to 15).
# Default levels are provided by the \code{mpm.levels} object.

v <- set.missing(u, ref = mpm.us[, 2:15], levels = mpm.levels)
print(v)

}
\seealso{
See \code{\link[morphonode]{set.rfcdata}} for random forest 
(RFC) data preparation and \code{\link[morphonode]{set.rbmdata}} for 
robust binomial model (RBM) data preparation.
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
