% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMcore.R
\name{set.rbmdata}
\alias{set.rbmdata}
\title{Ultrasound profile preparation for metastatic risk estimation}
\usage{
set.rbmdata(
  u,
  levels,
  short = 8,
  cortical = 2,
  ist = 1,
  ecs = 1,
  hab = 0,
  eco = 1,
  vp = c(1, 2, 3),
  vfl = c(2, 3, 4),
  ct = c(2),
  fid = c(1, 2, 3),
  cmid = c(2, 3, 4),
  shape = c(3),
  cs = 3,
  grouping = c(2, 3),
  ref = NULL,
  asFactor = FALSE,
  ...
)
}
\arguments{
\item{u}{New ultrasound profile generated by 
\code{\link[morphonode]{new.profile}}.}

\item{levels}{A list of length 14, corresponding to the levels of 
each ultrasound variable. Needed for categorical variables (factors); 
for continuous variables, it should assume the nominal value of 0. 
The default levels variable \code{\link[morphonode]{mpm.levels}} 
can be used.}

\item{short}{Numeric value corresponding to the short axis cutoff 
in millimeters (default = 8).}

\item{cortical}{Numeric value corresponding to the cortical thickness 
cutoff in millimeters (default = 2).}

\item{ist}{Dichotomous value {0, 1} for the presence of inflammatory 
stroma (perinodal hyperechogenic ring; default = 1).}

\item{ecs}{Dichotomous value {0, 1} for the presence of extracapsular 
spread (cortical interruption; default = 1).}

\item{hab}{Dichotomous value {0, 1} for the absence of the hilum 
(nodal core sign; default = 0).}

\item{eco}{Dichotomous value {0, 1} for heterogeneous echogenicity 
(echostructure; default = 1).}

\item{vp}{Categorical value (integers from 0 to 4) associated to a 
high-risk vascular flow architecture pattern (default = c(1, 2, 3)).}

\item{vfl}{Categorical value (integers from 0 to 4) associated to a 
high-risk vascular flow localization (default = c(2, 3, 4)).}

\item{ct}{Categorical value (integers from 0 to 4) associated to a 
high-risk cortical thickening (default = 2).}

\item{fid}{Categorical value (integers from 0 to 3) associated to a 
high-risk focal intranodal deposit (default = c(1, 2, 3)).}

\item{cmid}{Categorical value (integers from 0 to 4) associated to a 
high-risk cortical-medullar interface distortion (default = c(2, 3, 4)).}

\item{shape}{Categorical value (integers from 1 to 3) associated to a 
high-risk shape (default = 3).}

\item{cs}{Ordinal value (integers from 1 to 5) associated to a 
high-risk color score (default = 3).}

\item{grouping}{Categorical value (integers from 1 to 3) associated 
to a high-risk grouping (default = c(2, 3)).}

\item{ref}{Reference ultrasound features dataset as a (n, 14) 
data.frame, with n being the number of subjects (rows). 
The default simulated dataset \code{\link[morphonode]{mpm.us}} 
can be used.}

\item{asFactor}{Logical value. If TRUE, data.frame columns are converted 
to factors (default = FALSE).}

\item{...}{Currently ignored.}
}
\value{
A list of 2 objects:
\enumerate{
\item "ultrasound", dichotomized ultrasound features vector;
\item "missing", indices of missing values (empty if no missing 
   values are found).
}
}
\description{
Prepare a new ultrasound profile for metastatic risk 
   estimation using robust binomial modeling.
   This function includes missing values check and fix (see 
   \code{\link[morphonode]{set.missing}}).
}
\examples{

# Prepare a simulated malignant ultrasound profile
x <- new.profile(us.simulate(y = 1))
print(x)

# Set the new profile for RFC prediction
v <- set.rbmdata(x, ref = mpm.us[, 2:15], levels = mpm.levels)
print(v)

}
\references{
Fragomeni SM, Moro F, Palluzzi F, Mascilini F, Rufini V, Collarino A, 
Inzani F, Giacò L, Scambia G, Testa AC, Garganese G (2022). 
Evaluating the risk of inguinal lymph node metastases before surgery 
using the Morphonode Predictive Model: a prospective diagnostic study. 
Ultrasound xx Xxxxxxxxxx xxx Xxxxxxxxxx. 00(0):000-000.
<https://doi.org/00.0000/00000000000000000000>
}
\seealso{
See \code{\link[morphonode]{set.rfcdata}} for random forest  
   classifier data preparation. 
   See also \code{\link[morphonode]{us.predict}} to launch all 
   morphonode modules at once.
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
