% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMcore.R
\name{set.rfcdata}
\alias{set.rfcdata}
\title{Ultrasound profile preparation for random forest classification}
\usage{
set.rfcdata(u, levels = NULL, ref = NULL, ...)
}
\arguments{
\item{u}{New ultrasound profile generated by 
\code{\link[morphonode]{new.profile}}.}

\item{levels}{A list of length 14, corresponding to the levels of 
each ultrasound variable. Needed for categorical variables (factors); 
for continuous variables, it should assume the nominal value of 0. 
The default levels variable \code{\link[morphonode]{mpm.levels}} 
can be used.}

\item{ref}{Reference ultrasound features dataset as a (n, 14) 
data.frame, with n being the number of subjects (rows). 
The default simulated dataset \code{\link[morphonode]{mpm.us}} 
can be used.}

\item{...}{Currently ignored.}
}
\value{
A list of 2 objects:
\enumerate{
\item "ultrasound", ultrasound features vector;
\item "missing", indices of missing values (empty if no missing 
   values are found).
}
}
\description{
Prepare a new ultrasound profile for RFC prediction. 
   This function includes missing values check and fix (see 
   \code{\link[morphonode]{set.missing}}).
}
\examples{

# Prepare a simulated malignant ultrasound profile
x <- new.profile(us.simulate(y = 1))

# Set the new profile for RFC prediction
u <- set.rfcdata(x, ref = mpm.us[, 2:15], levels = mpm.levels)
print(u)

}
\references{
Fragomeni SM, Moro F, Palluzzi F, Mascilini F, Rufini V, Collarino A, 
Inzani F, Giacò L, Scambia G, Testa AC, Garganese G (2022). 
Evaluating the risk of inguinal lymph node metastases before surgery 
using the Morphonode Predictive Model: a prospective diagnostic study. 
Ultrasound xx Xxxxxxxxxx xxx Xxxxxxxxxx. 00(0):000-000.
<https://doi.org/00.0000/00000000000000000000>
}
\seealso{
See \code{\link[morphonode]{set.rbmdata}} for robust binomial 
   model data preparation. 
   See also \code{\link[morphonode]{us.predict}} to launch all 
   morphonode modules at once.
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
