% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMcore.R
\name{show.mpmenv}
\alias{show.mpmenv}
\title{Morphonode ultrasound variables description}
\usage{
show.mpmenv(brief = TRUE, ...)
}
\arguments{
\item{brief}{Logical value enabling a compact view (default = TRUE).}

\item{...}{Currently ignored.}
}
\description{
Shows a description of the utrasound variables used 
by the **morphonode** package.
}
\examples{
show.mpmenv()

}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
