% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMutils.R
\name{similarity}
\alias{similarity}
\title{Compute different similarity coefficients}
\usage{
similarity(x, y, f = "cosine", ceiling = 1e+07)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{f}{Similarity function. One among "cosine" (default), 
"jaccard" (for dichotomous vectors only), "pearson", "spearman",
"kendall", or "euclidean".}

\item{ceiling}{If \code{f = "euclidean"}, the similarity is 
computed as 1/distance. This argument limits the similarity to a 
very high value, in case the euclidean distance is equal to 0.
The default value is 1E+07.}

\item{...}{Currently ignored.}
}
\value{
A numeric value corresponding to the similarity coefficient.
}
\description{
Compute the similarity between two vectors.
}
\examples{

# Sample two random ultrasound profiles from the default dataset
x <- sample(mpm.us, 1, replace = FALSE, prob = NULL)
x <- as.matrix(x[, 2:15])
y <- sample(mpm.us, 1, replace = FALSE, prob = NULL)
y <- as.matrix(y[, 2:15])

# Compute the cosine similarity
r <- similarity(x, y)
print(r)

}
\references{
Leydesdorff L (2005). Similarity Measures, Author Cocitation 
Analysis,and Information Theory. In: JASIST 56(7), pp.769-772.
<https://doi.org/10.48550/arXiv.0911.4292>
}
\seealso{
\code{\link[morphonode]{jaccard}}, 
   \code{\link[morphonode]{euclidean}}
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
