% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMcore.R
\name{topsim}
\alias{topsim}
\title{Similarity search}
\usage{
topsim(
  v,
  x = mpm.us,
  f = "cosine",
  k = 5,
  p = 0.7,
  features = 2:15,
  check.data = TRUE,
  ...
)
}
\arguments{
\item{v}{An input vector of length equal to the number of columns 
of the reference data.frame (see below).}

\item{x}{Reference data.frame. If no reference is specified, 
the default simulated dataset (object \code{\link[morphonode]{mpm.us}}) 
will be used.}

\item{f}{Similarity function. Available functions: "cosine", "jaccard", 
"euclidean", "pearson", "spearman", "kendall" (default = "cosine").}

\item{k}{Numeric value defining the number of top-k profiles to 
return after similarity ranking (default = 5).}

\item{p}{Continuous value from 0 to 1 representing the minimum similarity 
value for an ultrasound profile to be included in the output 
(default = 0.7).}

\item{features}{Indices of the features (columns) in \code{x} used to 
compute similarity (default = 2:15).}

\item{check.data}{Logical value. If TRUE (default), the input data 
type is checked.}

\item{...}{Currently ignored.}
}
\value{
A list of 4 objects:
\enumerate{
\item "signature", metastatic risk signature (MRS);
\item "p", MRS-associated malignancy risk (evaluated as positive 
   predictive value, according to Fragomeni et al. 2022);
\item "ci95", 95% confidence intervals of p;
\item "y.uss", naive guess of the outcome (0: non-malignant, 
   1: malignant), based on the MRS (this will be less accurate than 
   the RFC-based prediction).
}
}
\description{
Vector similarity search across a reference dataset.
}
\examples{

# Prepare a simulated malignant ultrasound profile
x <- new.profile(us.simulate(y = 1))
u <- set.rfcdata(x, ref = mpm.us[, 2:15], levels = mpm.levels)

# Top-similar profiles
sim <- topsim(u$ultrasound)
print(sim)

}
\references{
Leydesdorff L (2005). Similarity Measures, Author Cocitation 
Analysis,and Information Theory. In: JASIST 56(7), pp.769-772.
<https://doi.org/10.48550/arXiv.0911.4292>
}
\seealso{
See \code{\link[morphonode]{us.predict}} to launch all 
   morphonode modules at once.
   See also \code{\link[morphonode]{ranksim}} for ultrasound profile 
   similarity ranking.
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
