% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMcore.R
\name{us.predict}
\alias{us.predict}
\title{Morhonode Predictive Model (MPM) launcher}
\usage{
us.predict(
  x,
  f = "cosine",
  levels = NULL,
  ref = NULL,
  rfc = NULL,
  rbm = NULL,
  k = 5,
  features = 2:15,
  orderbyDistance = FALSE,
  uncertainty = "loss",
  b0 = c(0, 0.028, 0.013),
  b = 0.055,
  rho = 0.9,
  wmax = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Ultrasound profile generated by the function 
\code{\link[morphonode]{new.profile}}).}

\item{f}{Similarity profiling core function: one between "cosine" 
(default) and "jaccard". The former directly compares ultrasound 
profiles, while the latter uses dichotomized versions of them 
(see also \code{\link[morphonode]{dichotomize}}).}

\item{levels}{A list of length 14, corresponding to the levels of 
each ultrasound variable. Needed for categorical variables (factors); 
for continuous variables, it should assume the nominal value of 0. 
If NULL (default), the internal \code{\link[morphonode]{mpm.levels}} 
variable will be used.}

\item{ref}{Reference ultrasound features dataset as a (n, 14) 
data.frame, with n being the number of subjects (rows). 
If NULL (default), the internal \code{\link[morphonode]{mpm.us}} 
variable will be used.}

\item{rfc}{Random forest classifier as an object of class 
\code{randomForest}. If NULL (default), the internal 
\code{\link[morphonode]{mpm.rfc}} variable will be used.}

\item{rbm}{Robust binomial model as a fitted model object of class 
\code{glm}. If NULL (default), the internal 
\code{\link[morphonode]{mpm.rbm}} variable will be used.}

\item{k}{Numeric value defining the number of top-k profiles to 
return after similarity ranking (default = 5).}

\item{features}{Indices of the features to be used for the similarity 
profiling (default = 2:15).}

\item{orderbyDistance}{Logical value enabling sorting by increasing 
euclidean distance of the top-similar ultrasound profiles 
(default = FALSE).}

\item{uncertainty}{Function used to compute the RFC prediction error. 
It can be one among "loss" (default) and "similarity". The former 
estimates the error on a new prediction based on a parametric linear 
relationship between the loss function and the observed (reference 
dataset) error. The latter estimates the error as the average of 
the observed errors of the top-3 similar profilesfrom the reference 
dataset (non-parametric).}

\item{b0}{Baseline uncertainty. A vector of three values representing 
the intercept parameter when \code{uncertainty = "loss"}. 
This value depends on the metastatic risk signature and it is 
fixed to 0 for the LMR one. The b0 values for MMR, HMR, and MET 
signatures are the first, second, and third element of the b0 
argument, respectively (default = c(0, 0.028, 0.013)).}

\item{b}{Uncertainty coefficient. A numeric value indicating the 
linear coefficient for the loss-to-error conversion equation.}

\item{rho}{Numeric value between 0 and 1 denoting the minimum required 
similarity coefficient to make a naive estimation of the true 
outcome y = {0, 1}. This guess is needed as a reference value for 
uncertainty calculation (default = 0.9).}

\item{wmax}{Nominal value for the self-correlation coefficient (for 
visualization purposes only; default = 1).}

\item{verbose}{If TRUE (default), a user-frienly summary of the 
prediction and estimation results is printed to screen.}

\item{...}{Currently ignored.}
}
\value{
A list of 5 objects:
\enumerate{
\item "prediction" (Morphonode-RFC module), a list including:
   \itemize{
       \item \code{y.hat}: the final malignancy prediction,
       \item \code{decisions}: the predictions of each RFC in the ensemble,
       \item \code{oob.err}: out-of-bag errors of each RFC in the ensemble;
    }
\item "E", estimated overall prediction error (Morphonode-RFC module);
\item "p", estimated malignancy risk(Morphonode-RBM module);
\item "signature", metastatic risk signature (Morphonode-DT module);
\item "profiles", data.frame containing the top-k similar profiles 
   sorted by similarity (Morphonode-SP module).
   This data.frame includes:
   \itemize{
       \item ID: numeric value identifying a subject,
       \item the 14 ultrasound features characterizing each subject,
       \item y: the observed outcome,
       \item E: subject-level estimated prediction error (Brier score),
       \item R: similarity coefficient with the input profile,
       \item D: euclidean distance from the input profile.
    }
}
}
\description{
The \code{us.predict} function launches the 4 MPM modules: 
   (i) malignancy prediction (Morphonode-RFC), (ii) malignancy risk 
   estimation (Morphonode-RBM), (iii) malignancy risk signature 
   detection (Morphonode-DT), and (iv) similarity profiling 
   (Morphonode-SP). 
   The MPM structure is described in Fragomeni et al. (2022). 
   See also the details section.
}
\details{
The MPM classifier (Morphonode-RFC module) is based on an 
   ensemble of 5 RFCs. Each RFC is trained over 10000 random trees, 
   with 3/14 randomly chosen variables per tree branching. The 5 RFCs 
   yield independent predictions and the majority wins. This module 
   provides a dichotomous phenotype classification in malignant (y = 1) 
   and non-malignant (y = 0), and an estimation of the prediction 
   error (E). Similarly, the Morphonode-RBM module provides a 
   continuous estimate of malignancy (i.e., p = malignancy risk), 
   through a binomial model with robust bootstrap standard error 
   estimation (5000 bootstrap iterations). Optimal cutpoint estimations 
   define two thresholds for p (three risk intervals): low risk (p < 0.23), 
   moderate risk (0.23 <= p <= 0.29), and high risk (p > 0.29). 
   In addition, Morphonode-DT model defines four metastatic risk 
   signatures, strongly associated with the metastasis rate in the 
   corresponding subjects. LMR (low metastatic risk) and MMR (moderate 
   metastatic risk) signatures are associated with none-to-low metastasis 
   rates. Conversely, HMR (high metastatic risk) and MET (metastatic) 
   signatures are associated with a high risk of single and multiple 
   metastatic events (lymph nodes), respectively. 
   Finally, the Morphonode-SP module ranks ultrasound profiles from 
   the reference dataset (by default, the internal simulated dataset) 
   by similarity with respect to the input profile. This provides a 
   supplementary support to the classification process, having only a 
   secondary role compared to the other three modules. Generally, 
   the majority of similar profiles should have the same outcome (y) 
   as the input one.
}
\examples{

# Create a simulated malignant ultrasound profile
x <- new.profile(us.simulate(y = 1))

# Lauch the Morhonode Predictive Model
u <- us.predict(x)

}
\references{
Fragomeni SM, Moro F, Palluzzi F, Mascilini F, Rufini V, Collarino A, 
Inzani F, Giacò L, Scambia G, Testa AC, Garganese G (2022). 
Evaluating the risk of inguinal lymph node metastases before surgery 
using the Morphonode Predictive Model: a prospective diagnostic study. 
Ultrasound xx Xxxxxxxxxx xxx Xxxxxxxxxx. 00(0):000-000.
<https://doi.org/00.0000/00000000000000000000>

Liaw A, Wiener M. Classification and Regression by randomForest (2002). 
R News, 2(3):18-22. <https://doi.org/10.1023/A:1010933404324>
}
\seealso{
See \code{\link[morphonode]{new.profile}} to create a new 
   ultrasound profile. 
   See also \code{\link[morphonode]{us.simulate}} for ultrasound 
   data simulation.
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
