% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMutils.R
\name{us.simulate}
\alias{us.simulate}
\title{Ultrasound profile simulation}
\usage{
us.simulate(
  reps = 1,
  x = mpm.us,
  y = NULL,
  signature = NULL,
  features = 2:15,
  header = NULL,
  ...
)
}
\arguments{
\item{reps}{Number of simulated vectors to be generated (default = 1).}

\item{x}{Reference ultrasound features data.frame. By default, a 
dataset of 948 simulated profiles (object \code{mpm.us}) is used.}

\item{y}{An optional numeric value defining the phenotype to be 
simulated: 0 for non-malignant and 1 for malignant (default = NULL).}

\item{signature}{An optional string defining the metastatic risk 
signature to be simulated. One among: "LMR" (low metastatic risk), 
"MMR" (moderate metastatic risk), "HMR" (high metastatic risk), 
"MET" (metastatic signature). Defaults to NULL.}

\item{features}{A vector of indices defining the ultasound features 
within the reference data.frame (default = 2:15).}

\item{header}{A vector of new names for the ultrasound features vector. 
If NULL (default), feature names will be imported from the reference.}

\item{...}{Currently ignored.}
}
\value{
A simulated ultrasound features vector.
}
\description{
Ultrasound profile simulation based on a reference dataset.
}
\examples{

# Simulate a generic ultrasound profile
u <- us.simulate()
print(u)

# Simulate a non-malignant ultrasound profile
u0 <- us.simulate(y = 0)
print(u0)

# Simulate a malignant ultrasound profile
u1 <- us.simulate(y = 1)
print(u1)

# Simulate a high metastatic risk profile (single metastasis event)
u.hmr <- us.simulate(signature = "HMR")
print(u.hmr)

# Simulate a metastatic ultrasound profile (multiple metastasis events)
u.met <- us.simulate(signature = "MET")
print(u.met)

}
\seealso{
See \code{\link[morphonode]{new.profile}} for ultrasound 
   profile creation and \code{\link[morphonode]{us.predict}} for 
   subject classification.
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
