% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMcore.R
\name{uss}
\alias{uss}
\title{Ultrasound signatures (uss) detection}
\usage{
uss(
  x,
  dichotomous = FALSE,
  dct = 2,
  short = 8,
  cortical = 2,
  ist = 1,
  ecs = 1,
  hab = 0,
  eco = 1,
  vp = c(1, 2, 3),
  vfl = c(2, 3, 4),
  ct = c(2),
  fid = c(1, 2, 3),
  cmid = c(2, 3, 4),
  shape = c(3),
  cs = 3,
  grouping = c(2, 3),
  ...
)
}
\arguments{
\item{x}{An (n, 14) ultrasound features data.frame, where n is the 
number of subjects.}

\item{dichotomous}{Logical value. The first step of this function is 
to dichotomize the input data.frame. Set \code{dichotomous = TRUE} 
if the input data.frame is already dichotomous (default = FALSE).}

\item{dct}{Numeric value. Moderate risk signature predicts a baseline 
risk of malignancy equal to 0.16 (CI95%: 0.06-0.25). According to 
Fragomeni et al. (2022), if at least 2 other ultrasound features 
than those used in the decision tree (referred to as diagnostic 
covariates) are above their optimal threshold, the malignancy risk 
increases to 0.55 (CI95%: 0.46-0.64). This signature will be marked 
as MMR1, in contrast to the basal MMR signature (MMR0).
The argument \code{dct} controls the number of diagnostic covariates 
needed to switch from MMR0 to MMR1 (default = 2).}

\item{short}{Numeric value corresponding to the short axis cutoff 
in millimeters (default = 8).}

\item{cortical}{Numeric value corresponding to the cortical thickness 
cutoff in millimeters (default = 2).}

\item{ist}{Dichotomous value {0, 1} for the presence of inflammatory 
stroma (perinodal hyperechogenic ring; default = 1).}

\item{ecs}{Dichotomous value {0, 1} for the presence of extracapsular 
spread (cortical interruption; default = 1).}

\item{hab}{Dichotomous value {0, 1} for the absence of the hilum 
(nodal core sign; default = 0).}

\item{eco}{Dichotomous value {0, 1} for heterogeneous echogenicity 
(echostructure; default = 1).}

\item{vp}{Categorical value (integers from 0 to 4) associated to a 
high-risk vascular flow architecture pattern (default = c(1, 2, 3)).}

\item{vfl}{Categorical value (integers from 0 to 4) associated to a 
high-risk vascular flow localization (default = c(2, 3, 4)).}

\item{ct}{Categorical value (integers from 0 to 4) associated to a 
high-risk cortical thickening (default = 2).}

\item{fid}{Categorical value (integers from 0 to 3) associated to a 
high-risk focal intranodal deposit (default = c(1, 2, 3)).}

\item{cmid}{Categorical value (integers from 0 to 4) associated to a 
high-risk cortical-medullar interface distortion (default = c(2, 3, 4)).}

\item{shape}{Categorical value (integers from 1 to 3) associated to a 
high-risk shape (default = 3).}

\item{cs}{Ordinal value (integers from 1 to 5) associated to a 
high-risk color score (default = 3).}

\item{grouping}{Categorical value (integers from 1 to 3) associated 
to a high-risk grouping (default = c(2, 3)).}

\item{...}{Currently ignored.}
}
\value{
A list of 4 objects:
\enumerate{
\item "signature", metastatic risk signature (MRS);
\item "p", MRS-associated malignancy risk (evaluated as positive 
   predictive value, according to Fragomeni et al. 2022);
\item "ci95", 95% confidence intervals of p;
\item "y.uss", naive guess of the outcome (0: non-malignant, 
   1: malignant) based on the MRS (this will be less accurate than 
   the RFC-based prediction).
}
}
\description{
This function implements the decision tree described in 
   Fragomeni et al. 2022 (Morphonode-DT module) to detect metastatic 
   risk signatures (MRSs). 
   The first split detects subjects with a "MET" (metastatic) signature. 
   These individuals show at least one among the three metastatic 
   markers (see details) and a high risk of malignancy (86-100%), 
   usually coming with multiple metastatic lymph nodes. 
   The other decision tree branches are defined on the base of five 
   key ultrasound features (see details), with a malignancy risk (MR)
   signature ranging from low ("LMR", 0-10%) to moderate ("MMR", 6-25%) 
   to high ("HMR", 52-90%), implying a single metastatic lymph node 
   in most malignancies.
   The main goal of MRSs is to predict single-metastatic event 
   malignancies (HMR) and multiple-metastatic event malignancies (MET).
}
\details{
The core method of the Morphonode-DT model is implemented 
   in this function. A series of binary branching points define the 
   metastatic risk signature (MRS) of the subject.
   The first branch point is based on the evaluation of three metastatic 
   markers: the absence of the nodal core sign (hilum), the presence 
   of the perinodal hyperechogenic ring, and the presence of cortical 
   interruption. If at least one of these conditions are true, the 
   ultrasound profile has a high malignancy risk (86-100%) with 
   multiple metastatic lymph nodes (Fragomeni et al. 2022). This is 
   referred to as the metastatic (MET) signature.
   A cortical thickness below 2 mm defines a low metastatic risk (LMR) 
   signature (0.04, CI95%: 0.00-0.10). LMR subjects are mostly mot 
   malignant. 
   Based on the values of short axis, vascular flow architecture pattern, 
   cortical thickening, and vascular flow localization, two more 
   signatures are defined: (i) moderate metastatic risk (MMR; 0.16, 
   CI95%: 0.06-0.25), and (ii) high metastatic risk (HMR; 0.81, 
   CI95%: 0.52-0.90). Similarly to the MET signature, HMR subjects are 
   mostly malignant, but chraracterized by a single metastatic event.
   MRSs should be always compared to the output of the random forest 
   classifier (Morphonode-RFC module) and robust binomial model 
   (Morphonode-RBM module). 
   The main advantage of MRSs is the prediction of either multiple 
   (MET signature) or single (HMR signature) metastasis events.
}
\examples{

# Extract 5 random subjects from the default simulated dataset
x <- mosaic::sample(mpm.us[, 2:15], 5, replace = FALSE, prob = NULL)
print(x)

# Assign a metastatic risk signature to each subject in the dataset
mrs <- uss(x)
x$signature <- mrs$signature
print(x)

}
\references{
Fragomeni SM, Moro F, Palluzzi F, Mascilini F, Rufini V, Collarino A, 
Inzani F, Giacò L, Scambia G, Testa AC, Garganese G (2022). 
Evaluating the risk of inguinal lymph node metastases before surgery 
using the Morphonode Predictive Model: a prospective diagnostic study. 
Ultrasound xx Xxxxxxxxxx xxx Xxxxxxxxxx. 00(0):000-000.
<https://doi.org/00.0000/00000000000000000000>
}
\seealso{
See \code{\link[morphonode]{us.predict}} to launch all 
   morphonode modules at once.
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
