% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMutils.R
\name{vpart}
\alias{vpart}
\title{Data partitioning utility}
\usage{
vpart(data, p, shuffle = FALSE, ...)
}
\arguments{
\item{data}{An input data.frame.}

\item{p}{Proportion of rows to be extracted from the input dataset.}

\item{shuffle}{A logical value. If TRUE, the input rows are randomly 
shuffled before data partitioning.}

\item{...}{Currently ignored.}
}
\value{
A list of 2 data.frames:
\enumerate{
\item "training.set", the portion of the input data defined by p;
\item "validation.set", the portion of the input data defined by 1-p.
}
}
\description{
Extract a random partition from an input dataset.
}
\examples{

# Extract a subset of 300 subjects and an outcome vector of length 30 
# from the default simulated dataset

x <- mosaic::sample(mpm.us, 300, replace = FALSE, prob = NULL)

# Data partitioning

x <- vpart(x, p = 0.75)
print(dim(x$training.set))
print(dim(x$validation.set))

}
\seealso{
\code{\link[morphonode]{buildPredictor}},
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
